refactor(collection): move parseTips to collection package and add edge case test
parseTips() belongs to the parsing/business logic layer, not the UI file.
Moved to CollectionViewModel.kt as a package-level internal function.
Added regression test for dash-without-space (no new bullet started).
