+++ vertexai/app/src/main/kotlin/com/formulae/chef/feature/collection/CollectionViewModel.kt
@@ -321,3 +321,25 @@ class CollectionViewModel(
         const val MAX_SERVINGS = 30
     }
 }
+
+internal fun parseTips(tipsAndTricks: String): List<String> {
+    val lines = tipsAndTricks.lines()
+    val tips = mutableListOf<String>()
+    val current = StringBuilder()
+    for (line in lines) {
+        when {
+            line.startsWith("- ") -> {
+                if (current.isNotEmpty()) tips.add(current.toString().trim())
+                current.clear()
+                current.append(line.removePrefix("- ").trim())
+            }
+            line.isNotBlank() -> {
+                if (current.isNotEmpty()) current.append(" ")
+                current.append(line.trim())
+            }
+        }
+    }
+    if (current.isNotEmpty()) tips.add(current.toString().trim())
+    if (tips.isEmpty() && tipsAndTricks.isNotBlank()) tips.add(tipsAndTricks.trim())
+    return tips
+}
